+++ HISTORY.rst
@@ -210,3 +210,6 @@ History
 * `filtered` has been update to use a list of filters `AppliedFilter` objects and can include stage.
 * `Comments` is now an object with attributes including `value`, `author`, and `date`.
 * Moved many common class objects to folder called `common` to reduce redundancy.
+* In `timeseries.Run` use `add_channel` to add a channel if you `append`, `extend`, or `insert` a channel, then you must run `Run._update_channel` to update the metata.
+* `MetadataBase` no longer overrides `__deepcopy__` uses `model_copy(deep=True)` under the hood.
+* Cannot use len() on `Run` or `Channel` objects, use `len(Run.channels)` or `len(Channel)` instead.